refactoring, modular setup
